TA: Fixed issue of the fw parser. [1/1]
PD#SWPL-24152

Problem:
The picture is black after playing mpeg2 bitstream or video.

Solution:
Due to the existence of invalid firmware in the firmware table,
when processing dup, the firmware analysis error is caused by
the valid firmware. In order to solve this problem, the invalid
fw needs to be removed from table, and this action needs to be
processed when filtering the valid fw to avoid the possibility
of overflow of the firmware table.

Verify:
TM2

Change-Id: I36f7d416f9452a2c45cdff98d251362a32e79a89
Signed-off-by: Nanxin Qin <nanxin.qin@amlogic.com>